+++ src/main/java/com/example/locadora/services/ClasseService.java
@@ -36,12 +36,19 @@ public class ClasseService {
         }
     }
 
-    public void deletar(Long id) {
+    public void deletar(Long id) throws Exception{
 
         Optional<Classe> classeExistente = classeRepository.findById(id);
 
         if (classeExistente.isPresent()) {
-            classeRepository.delete(classeExistente.get());
+
+            if(classeExistente.get().getTitulos().isEmpty()){
+                classeRepository.delete(classeExistente.get());
+            } else {
+                throw new Exception("Classe está relacionada a um ou mais títulos e não pode ser excluída.");
+            }
+
+
         } else {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Classe não encontrada");
         }
